# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DOCS",
        "description": "Advanced API documentation",
        "contact": {
            "email": "narushevich.maksim@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API documentation"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Register new user and return JWT token",
                "description": "Create a new user.",
                "operationId": "addUser",
                "requestBody": {
                    "description": "Create user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewUser"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Null response"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            }
        },
        "/confirm": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Confirm registration",
                "description": "Returns successful confirmation information",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::confirm",
                "requestBody": {
                    "description": "Confirmation token",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login with existing user and obtain JWT token",
                "description": "Login user.",
                "operationId": "loginUser",
                "requestBody": {
                    "description": "Authorize user and get token",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Null response"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            }
        },
        "/auth-user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get authorized user details",
                "description": "Returns logged user data",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::getUser",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserItem"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "All users",
                "description": "Returns liss of all users",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::getUserList",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserItem"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/delete/{id}": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete user by ID",
                "description": "Returns successful delete information",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::deleteUser",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth-user/delete": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete authorized user",
                "description": "Returns successful delete information",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::deleteAuthUser",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "NewUser": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/User"
                    },
                    {
                        "required": [
                            "name",
                            "c_password"
                        ],
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "c_password": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "UserItem": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/User"
                    },
                    {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "name": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string"
                            },
                            "email_verified_at": {
                                "type": "string",
                                "format": "date"
                            },
                            "created_at": {
                                "type": "string",
                                "format": "date"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "date"
                            }
                        },
                        "type": "object"
                    }
                ]
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Oauth2 security",
                "name": "oauth2",
                "in": "header",
                "flows": {
                    "clientCredentials": {
                        "tokenUrl": "oauth/token",
                        "scopes": []
                    }
                },
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}